Engage-MVC now adds controller and views to your project.

+++ packaging/packaging.rb
@@ -14,10 +14,24 @@ task :prepare_package_core do
 end
 
 task :prepare_package_mvc do
-  output_directory_mvc = './packaging/EngageNet-Mvc/lib/40/'
-  FileUtils.mkdir_p output_directory_mvc
+  output_directory_mvc_lib = './packaging/EngageNet-Mvc/lib/40/'
+  output_directory_mvc_content_controllers = './packaging/EngageNet-Mvc/content/Controllers/'
+  output_directory_mvc_content_views = './packaging/EngageNet-Mvc/content/Views/Engage/'
+  FileUtils.mkdir_p output_directory_mvc_lib
+  FileUtils.mkdir_p output_directory_mvc_content_controllers
+  FileUtils.mkdir_p output_directory_mvc_content_views
 
-  copy_files './src/EngageNet.Mvc/bin/Release/', output_directory_mvc, 'EngageNet.Mvc', ['dll', 'pdb', 'xml']
+  copy_files './src/EngageNet.Mvc/bin/Release/', output_directory_mvc_lib, 'EngageNet.Mvc', ['dll', 'pdb', 'xml']
+  FileUtils.cp './src/EngageNet.SampleWebsiteMvc2/Controllers/EngageController.cs',
+               output_directory_mvc_content_controllers + 'EngageController.cs.pp'
+  copy_files './src/EngageNet.SampleWebsiteMvc2/Views/Engage/', output_directory_mvc_content_views, 'LogOn', ['aspx']
+  copy_files './src/EngageNet.SampleWebsiteMvc2/Views/Engage/', output_directory_mvc_content_views, 'LogOnCancelled', ['aspx']
+  copy_files './src/EngageNet.SampleWebsiteMvc2/Views/Engage/', output_directory_mvc_content_views, 'LogOnSuccess', ['aspx']
+  
+  text = File.read output_directory_mvc_content_controllers + 'EngageController.cs.pp'
+  File.open output_directory_mvc_content_controllers + 'EngageController.cs.pp', 'w' do |file|
+  	file.puts text.gsub /EngageNet\.SampleWebsiteMvc2/, '$rootnamespace$'
+  end
 end
 
 exec :package_core => :prepare_package_core do |cmd|
@@ -43,8 +57,10 @@ task :package => [:package_core, :package_mvc] do
 end
 
 task :clean_packages do
-	FileUtils.rm_r './packaging/EngageNet/lib/'
+	FileUtils.rm_r './packaging/EngageNet/lib/' unless not File.directory? './packaging/EngageNet/lib/'
 	FileUtils.rm Dir.glob './packaging/EngageNet/*.nupkg'
-	FileUtils.rm_r './packaging/EngageNet-Mvc/lib/'
+
+	FileUtils.rm_r './packaging/EngageNet-Mvc/lib/' unless not File.directory? './packaging/EngageNet-Mvc/lib/'
+	FileUtils.rm_r './packaging/EngageNet-Mvc/content/' unless not File.directory? './packaging/EngageNet-Mvc/content/'
 	FileUtils.rm Dir.glob './packaging/EngageNet-Mvc/*.nupkg'
 end
